added user validation

+++ app/Http/Controllers/PostsController.php
@@ -9,6 +9,16 @@ use App\User;
 class PostsController extends Controller
 {
     /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except'=>['index', 'show']]);
+    }
+
+    /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
@@ -72,6 +82,10 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+        //Check for correct user
+        if (auth()->user()->id!==$post->user_id) {
+            return redirect('posts')->with('error', 'You do not have permission to do this action!');
+        }
         return view('posts.edit')->with('post', $post);
     }
 
@@ -106,6 +120,10 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        //Check for correct user
+        if (auth()->user()->id!==$post->user_id) {
+            return redirect('posts')->with('error', 'You do not have permission to do this action!');
+        }
         $post->delete();
         return redirect('/posts')->with('success', 'Post Deleted!');
     }
